restructured project for nextjs app router
