+++ frontend/src/app/layout.js
@@ -1,6 +1,7 @@
 import './globals.css'
 import { Inter } from 'next/font/google'
-import AuthProvider from './auth/AuthProvider'
+import AuthProvider from '../auth/AuthProvider'
+import NavBar from '@/components/navbar';
 
 const inter = Inter({ subsets: ['latin'] })
 
@@ -13,7 +14,9 @@ export default function RootLayout({ children }) {
   return (
     <html lang="en">
       <body className={inter.className}>
+        <h1 class="text-4xl font-bold text-center my-8">Pikadle</h1>
         <AuthProvider>
+          <NavBar />
           {children}
         </AuthProvider>
       </body>
