+++ frontend/src/app/login/page.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import LoginForm from './LoginForm'
+import LoginForm from '@/auth/LoginForm'
 
 
 const Login = () => {
