+++ frontend/src/app/register/page.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import RegisterForm from '@/auth/RegisterForm';
+
+const Register = () => {
+    return (
+        <RegisterForm />
+    )
+}
+
+export default Register;
