+++ frontend/src/app/safari/page.js
@@ -2,7 +2,7 @@
 //Safari Idea: 20 guesses to catch as many pokemon as possible
 //Each pokemon will be a certain type
 import { useEffect, useState, useMemo, useRef } from "react"
-import Guesses from './guesses.js'
+import Guesses from '../../components/guesses.js'
 import { Typeahead } from 'react-bootstrap-typeahead';
 import Form from 'react-bootstrap/Form';
 import './wordle.css'
