+++ frontend/src/auth/LoginForm.jsx
@@ -1,8 +1,11 @@
 "use client"
 
 import { signIn } from 'next-auth/react';
+import { useRouter } from 'next/navigation';
 
 export default function Form() {
+    const router = useRouter();
+
     const handleSubmit = async (event) => {
         event.preventDefault();
         const data = new FormData(event.currentTarget);
@@ -24,7 +27,7 @@ export default function Form() {
                 <input type="text" name="username" />
                 <input type="password" name="password" />
                 <button type="submit">Sign in</button>
-                <button>Sign up</button>
+                <button type="button" onClick={() => router.push("/register")}>Sign up</button>
             </form>
         </>
     )
