+++ frontend/src/auth/RegisterForm.jsx
@@ -1,3 +1,5 @@
+"use client";
+
 import React from 'react'
 import { register } from './requests'
